requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
altair==4.2.2
anyio==3.6.2
async-timeout==4.0.2
asynctest==0.13.0
attrs==22.2.0
backports.zoneinfo==0.2.1
bcrypt==4.0.1
biopython==1.81
blinker==1.6
cachetools==5.3.0
certifi==2022.12.7
charset-normalizer==3.1.0
click==8.1.3
cycler==0.11.0
datasets==2.11.0
decorator==5.1.1
dill==0.3.6
dnspython==2.3.0
email-validator==1.3.1
entrypoints==0.4
fastapi==0.95.0
filelock==3.11.0
fonttools==4.38.0
frozenlist==1.3.3
fsspec==2023.1.0
gitdb==4.0.10
GitPython==3.1.31
h11==0.14.0
httpcore==0.16.3
httptools==0.5.0
httpx==0.23.3
huggingface-hub==0.13.4
idna==3.4
importlib-metadata==6.1.0
importlib-resources==5.12.0
itsdangerous==2.1.2
Jinja2==3.1.2
jsonschema==4.17.3
kiwisolver==1.4.4
loguru==0.7.0
markdown-it-py==2.2.0
MarkupSafe==2.1.2
matplotlib==3.5.3
mdurl==0.1.2
mock==1.0.1
multidict==6.0.4
multiprocess==0.70.14
numpy==1.21.6
openai==0.8.0
orjson==3.8.9
packaging==23.0
pandas==1.1.5
Pillow==9.5.0
pinecone-client==2.2.1
pkgutil_resolve_name==1.3.10
protobuf==3.20.3
pyarrow==11.0.0
pydantic==1.10.7
pydeck==0.8.0
Pygments==2.14.0
PyJWT==2.6.0
Pympler==1.0.1
pyparsing==3.0.9
pyrsistent==0.19.3
pysam==0.20.0
python-dateutil==2.8.2
python-decouple==3.8
python-dotenv==0.21.1
python-multipart==0.0.6
pytz==2023.3
pytz-deprecation-shim==0.1.0.post0
PyYAML==6.0
requests==2.28.2
responses==0.18.0
rfc3986==1.5.0
rich==13.3.3
semver==3.0.0
six==1.16.0
smmap==5.0.0
sniffio==1.3.0
starlette==0.26.1
streamlit==1.20.0
toml==0.10.2
toolz==0.12.0
tornado==6.2
tqdm==4.65.0
typing_extensions==4.5.0
tzdata==2023.3
tzlocal==4.3
ujson==5.7.0
urllib3==1.26.15
uvicorn==0.21.1
uvloop==0.17.0
validators==0.20.0
watchfiles==0.19.0
websockets==11.0.1
xxhash==3.2.0
yarl==1.8.2
zipp==3.15.0
